fixup! 2.17: changed mob's experience from 32-bit to 64-bit value

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
@@ -834,7 +834,7 @@ public abstract class Mob extends Char {
 		}
 
 		if (buff(Longsword.HolyExpEffect.class) != null){
-			EXP = (int) Math.round(EXP * Math.pow(1.15f, buff(Longsword.HolyExpEffect.class).stacks));
+			EXP = Math.round(EXP * Math.pow(1.15f, buff(Longsword.HolyExpEffect.class).stacks));
 		}
 
 		if (this.buff(Overload.class) != null) {
